new unregistering mechanism
ghc-package.eclass was too lazy to unregister packages.
It only checked if another version was still available,
and if there was, it wouldn't unregister any package,
assuming that the other version would provide exactly
the same Haskell packages than the removed one.

That was already wrong before, but unlikely to cause
trouble, because it occurred relatively rarely that
Haskell package names changed between releases.

With the advent of version numbers, this is however
no longer true. The Haskell package names do almost
always change. The new code is not beautiful, but at
least designed to do the right thing. The change should
be backwards-compatible, but if anything breaks, please
let me know.