strict-concurrency-fix-and-ebuild
in version 0.2.2 parallel is set to any, but with parallel-3.x it will failed to build, so if both versions
installed it will choose parallel-3.x before parallel-2.x. thats the patch about

the new ebuild won't break happstack-util, and the other dependency vacuum-cairo fails to build with gtk it seems.
(but there is also the problem with parallel-3 and parallel-2 but this is not the only fix there have to be done, if something
would be done)