Enable gtk flag for criterion if chart enabled.
Chart depends on gtk itself, so it shouldn't be necessary to add an
additional useflag for gtk. If chart is enabled, then the Gtk window
should be enabled as there are no additional dependencies. Currently, it
is impossible to enable Gtk Windows because the ebuild never enables the
criterion cabal flag `gtk` (which doesn't actually add any additional
dependencies).