Add conduit-0.3 [masked]
Conduit 0.3 shouldn't be "released" yet because currently it breaks
Yesod, a major user of conduit. This adds most of the conduit libraries,
updated for conduit-0.3, hardmasked. Notable exceptions to the
hardmasking are "resourcet" and "zlib-bindings" which are simply other
dependencies of these packages which needed to be updated.

Not all of these are tested. I've tested conduit and resourcet only, the
others are added for convenience as simply hackported ebuilds.