Always fetch app-emacs/haskell-mode-9999 via HTTPS
This commit changes the `EGIT_REPO_URI` of
`app-emacs/haskell-mode-9999` to always use the HTTPS Git transport
protocol, rather than either that or the plain Git transport protocol.

The only reasonable scenario I see in which the plain Git transport
would work and the HTTPS transport wouldn’t is that of HTTPS being
blocked by an attacker wishing to force the download to occur via the
unsecured plain Git transport, enabling the injection of malware.