app-emacs/haskell-mode: generate autoloads in src_compile.
The haskell-mode-autoloads.el file used to be part of the releases,
and it is needed by our 50haskell-mode-gentoo.el. However, in newer
versions of haskell-mode, that file must be built.

This new revision adds an src_compile function which prepends a single
command to the default elisp_src_compile, namely

  emake haskell-mode-autoloads.el

That is sufficient to build the autoloads file, and avoids
precompiling any of the other elisp files (which will be compiled by
the elisp eclass anyway).

Package-Manager: portage-2.3.0