eclass/ghc-package.eclass: fix ghc-install-pkg failure to register
Prior to this patch, ghc-install-pkg dies when installing
dev-haskell/haddock-library-1.6.0. The function dies trying to read
from stdin with the getContents function, called in the registerPackage
function in ghc-pkg, because it tries to read a directory rather than a file.

This patch changes the behaviour of the ghc-install-pkg function such
that instead of pushing ${pkg_config_file} onto stdin and instructing
the registerPackage function to call `hGetContents stdin`, ghc-install-pkg
now pushes ${pkg_config_file} directly into ghc-pkg's registerPackage
function, resulting in the `readutf8file` function being called directly on
${pkg_config_file}.