dev-lang/ghc: don't strip ${CTARGET} libraries
Posting the comment note here as-is:
"""
When we build a cross-compiler the layout is the following:
    usr/lib/${CTARGET}-ghc-${VER}/ contains target libraries
but
    usr/lib/${CTARGET}-ghc-${VER}/bin/ directory
containst host binaries (modulo bugs).

Portage's stripping mechanism does not skip stripping
foreign binaries. This frequently causes binaries to be
broken.

Thus below we disable stripping of target libraries and allow
stripping hosts executables.
"""

While at it update to EAPI=7.

Reported-by: Shane Peelar
Closes: https://github.com/gentoo-haskell/gentoo-haskell/issues/960
Package-Manager: Portage-2.3.73, Repoman-2.3.17
Signed-off-by: Sergei Trofimovich <slyfox@gentoo.org>