dev-haskell/lhs2tex: bump up to 1.24
NOTE: I do not know if this package works as intended at runtime,
as I am unfamiliar with it. But it does build.

The previously available version does not build with ghc-8.6.5,
and we needed to adjust the dependencies anyway due to the ::gentoo
masking of dev-tex/polytable.

Package-Manager: Portage-2.3.103, Repoman-2.3.23
Signed-off-by: Jack Todaro <solpeth@posteo.org>