travis: remove travis-ci configuration
config files have been added for github actions to execute both pkgcheck
and repoman.

the current travos-ci config fails the pkgcheck stage due to some sort
of python issues with portage. this was not easily resolved by bumping
the python version.

due to this, and the fact that travis seems to be moving away from a
free model, we're going to just stop using travis altogether

Signed-off-by: Wolfgang E. Sanyer <WolfgangESanyer@gmail.com>