dev-haskell/haddock: remove old live ebuild
It looks like it was first added back when v2.11.0 was the latest
release, in june 2012. based on git history, the live ebuild was needed
in order to pull in pre-release patches.

Since we're up to v2.24.0, and I don't see any compelling non-released
upstream patches, and this ebuild seems undermaintained, and it is the
only package that still depends on the older version of haddock-api...

I'm dropping it.

Signed-off-by: Wolfgang E. Sanyer <WolfgangESanyer@gmail.com>