eclass/haskell-cabal.eclass: Add support for Hackage revisions (#1275)
This allows for automatic downloading and extracting of cabal revisions from
Hackage. It also uses the `dos2unix` utility to convert any CRLF-encoded cabal
files, which will help facilitate patching in the future.

This functionality can be enabled by:
1. Adding `CABAL_HACKAGE_REVISION=` with the revision number above
   `inherit haskell-cabal`. (A value less than `1` has no effect).
2. Renaming any `MY_PN` definitions to `CABAL_PN` and moving them
   _above_ `inherit haskell-cabal`

If these steps are not taken then it will default to the original behavior, with
the exception that `dos2unix ${CABAL_PN}.cabal` is now run by default in
`src_prepare`, unless this function is overridden.

- `SRC_URI` is now set in the eclass, and is entirely optional unless using an
  alternate source for tarballs.
- `cabal_chdeps` can now be used without defining `src_prepare`, by using the
  `CABAL_CHDEPS` variable.
- `MY_PN` is now deprecated in favor of `CABAL_PN`