dev-lang/ghc: remove prebuilt libHScontainers from sources
At release, a prebuilt x86 binary for the libHScontainers library was
accidentally included in the source tarball.  This was noticed by the
Fedora packagers here:
https://www.mail-archive.com/search?l=ghc-devs@haskell.org&q=subject:%22Re%5C%3A+%5C%5BHaskell%5C%5D+%5C%5BANNOUNCE%5C%5D+GHC+9.0.2+released%22&o=newest&f=1
but no action was taken by upstream, so they are patching it out in
their build recipe:
https://src.fedoraproject.org/rpms/ghc9.0/blob/rawhide/f/ghc9.0.spec#_327
We should do the same.  From looking at upstream source releases, more
recent versions like the 9.2 branch do not have the same mistake, so
this is specific to 9.0.

Also fixes an incorrect hash that was somehow committed for the ppc64le
binary.

Closes: https://bugs.gentoo.org/863686
Closes: https://github.com/gentoo/gentoo/pull/26876
Signed-off-by: Sam James <sam@gentoo.org>
Signed-off-by: hololeap <hololeap@users.noreply.github.com>