haskell-cabal.eclass: Limit exposed packages for Setup.hs
The packages available when compiling Setup.hs needs to be controlled,
otherwise module name collisions are possible.

Uses the -hide-all-packages option for ghc which requires exposed
packages to be listed explicitly. The only packages that are exposed by
default are Cabal and base. Others can be added in the ebuild using the
new GHC_BOOTSTRAP_PACKAGES variable.

Bug: https://github.com/gentoo-haskell/gentoo-haskell/issues/1327
Signed-off-by: hololeap <hololeap@users.noreply.github.com>