haskell-cabal.eclass: Add cabal-register-inplace()
Adds a new function, cabal-register-inplace(), and inserts before
'./setup test' is run.

This function registers the package with the in-place pacakge DB,
located in "${S}/dist/package.conf.inplace/". This is sometimes needed
for tests when the package is not yet installed. Unfortunately, prebuilt
solutions to this problem, such as './setup register --inplace', do not
seem to work correctly.

Signed-off-by: hololeap <hololeap@users.noreply.github.com>