haskell-cabal.eclass: Fix for DOS encoding
The sed expression in cabal-register-inplace() is not resilient to DOS encoded
files and will result in errors if it encounters a '\r' character, as it will
add it to the version string.

(It does not seem to think an '\r' at the end is part of the newline.)

This adds an optional '\r' to the end of the sed regex to make sure it is
exluded if it is encountered.

Signed-off-by: hololeap <hololeap@users.noreply.github.com>