haskell-cabal.eclass: Use bash globbing
Using sed to find the package version has proven unreliable. This fix
uses the find utility with a globbed shell pattern for the `-name` flag
in order to find the *.conf file, so the version number does not need
to be determined.

In the future, we can use a Haskell program using the Cabal library to
reliably parse the version number and/or outright replace the
cabal-register-inplace function with a more programmatic solution.

Bug: https://github.com/gentoo-haskell/gentoo-haskell/issues/1270
Signed-off-by: hololeap <hololeap@users.noreply.github.com>