dev-haskell/process: Mask for removal
Exposing the 'process' package as a portage package has proven to be
a bad idea, creating many complications and system inconsistencies.

The version that comes bundled with GHC needs to be the only version
on the system. If upgrades to the 'process' package are needed, it
needs to be done through dev-lang/ghc.

Signed-off-by: hololeap <hololeap@users.noreply.github.com>