dev-haskell/cabal: drop 3.6.3.0-r1
This revision bump was a hack and is no longer needed as
dev-hackport/process is now masked and the 'process' package is
exclusively built through dev-lang/ghc.

Users may continue upgrading dev-haskell/cabal _or_ using the version
that comes bundled with dev-lang/ghc, as both now depend on the same
version of 'process'.

Signed-off-by: hololeap <hololeap@users.noreply.github.com>