*/*: Remove ~x86 (and related) from KEYWORDS
x86 is mostly untested, and even building a subset of ::haskell
revealed failing tests which pass on amd64.

With the new behavior of hackport to only add ARCH to KEYWORDS, there
will be CI failures if a package has ~x86 removed but not from its
revdep tree as well. The only sensible mitigation is to remove ~x86
tree-wide until we have a dedicated tester willing to fix current bugs.

See: https://github.com/snoyberg/mono-traversable/issues/210
Signed-off-by: hololeap <hololeap@users.noreply.github.com>