dev-haskell/yaml: make raw-strings-qq required test dep
In yaml's cabal file, raw-strings-qq is listed unconditionally
under the spec test-suite's dependency list. In our ebuild,
raw-strings-qq was listed as conditional to the examples USE flag.

In a default USE flag scenario, src_configure fails as unable to
find raw-strings-qq if not already coincidentally on the system.

This only affects FEATURES=test users with the default USE flag
setup.

Signed-off-by: Jack Todaro <solpeth@posteo.org>