dev-haskell/integer-logarithms: revbump to override ::gentoo
The reporter indicated a configure failure of this package after
enabling the ::haskell repo to access its broader range of packages.

Given that both ::gentoo and ::haskell contain an -r1 of this package,
it is suspected that Portage did not force the system to update to the
::haskell version of it (as they are technically the same version).

However, the ::gentoo version does not contain up-to-date hackage
revisions. This caused a configure failure for the reporter.

By rev-bumping this package, Portage should force the system to update
to the ::haskell version and successfully configure.

Reported-By: @alecStewart1
Closes: https://github.com/gentoo-haskell/gentoo-haskell/issues/1470
Signed-off-by: Jack Todaro <solpeth@posteo.org>