dev-haskell/*: Tighten deps for old HLS packages
The new haskell-language-server versions combine the various plugin
packages into the main HLS package. This means that hls-*-plugin
packages are outdated and will eventually be removed. By tightening the
dependencies on the old HLS packages, we ensure that they will not be
pulled into the dependency graph as subslot rebuilds when remaining HLS
packages are updated.

Signed-off-by: hololeap <hololeap@protonmail.com>