dev-lang/ghc: Use CHECKREQS_MEMORY for USE=doc
The version of haddock bundled with ghc-9.4.* uses 6-7GB of memory when
building the docs for GHC. This has caused sporadic failures even on a
system with 12GB of memory, so check for at least 16GB to be safe.

The versions of haddock bundled with >=ghc-9.6 are supposed to be much
less memory consumptive, but, unfortunately, it is quite difficult to
backport these newer haddock versions to ghc-9.4.

Signed-off-by: hololeap <hololeap@protonmail.com>