dev-lang/ghc: Delete buggy tests during src_prepare
Unfortunately, the --broken-test option for hadrian does not work as one
might expect: It still runs the test but then fails if the test
succeeds, which means buggy tests can still cause test suite failures.
Instead, we give the full path to the test and delete it during
src_prepare, which is a the most straightforward and reliable way to
disable the test completely.

Signed-off-by: hololeap <hololeap@protonmail.com>