dev-lang/ghc: Revert "Disable need for alex during build"
This fix is not working properly and the proper solution is still not
in an official cabal-install release. Building without USE=ghcbootstrap
will not be feasible until the --ignore-build-tools flag for cabal makes
its way into the hadrian-bootstrap-sources tarball.

In the meantime, install ghc-9.8 and then bootstrap ghc-9.10 using the
instructions found here:

https://github.com/gentoo-haskell/gentoo-haskell/wiki/Building-experimental-ghc#bootstrap-install

This reverts commit 27185cb8c4e4cfe6b3c59114d6970268d553cdbf.

See: https://github.com/haskell/cabal/pull/10128
Bug: https://github.com/gentoo-haskell/gentoo-haskell/issues/1552
Signed-off-by: hololeap <hololeap@protonmail.com>