Create pull_request_template.md
Copied verbatim from the ::gentoo repository.

As a Gentoo project repository we are also bound by the copyright and certificate of origin requirements. We should make that clear in PRs.

Sign-offs and pkgcheck scans can be automated using simple Portage configuration changes and pkgdev tooling for commits and pushes, so this is not a major imposition.

This also means there is a clear chain of accountability for copyright and Gentoo policy adherence for whenever commits are synced over to ::gentoo.