dev-haskell/cabal: loosen unix dep for 3.14.1.1
There is a new(ish) lower bound on "unix" because, apparently, some
earlier releases had i386/arm64 bugs. OK, but the new lower bound
excludes the version of unix that ships with ghc-9.10.1, so if you are
trying to bootstrap ghc-9.12 with 9.10 by upgrading hadrian et al., you
will hit a wall.

Signed-off-by: Michael Orlitzky <mjo@gentoo.org>