dev-util/ghcup-bin: avoid die() in global scope
The ghcup_bin_prefix() function can die on an unsupported arch, but it's
called in global scope to set some variables like

  GHCUP_BIN_PREFIX="$(ghcup_bin_prefix "${ARCH}")"

As a result the ebuild is un-sourceable on some arches, and that in turn
kills the metadata generation process. It's easy enough to fix by declaring
those global variables in pkg_setup().

Signed-off-by: Michael Orlitzky <mjo@gentoo.org>